# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AlexTa",
        "description": "Documentation de l'API de l'application AlexTa",
        "contact": {
            "email": "support@alexta.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Lister tous les utilisateurs",
                "operationId": "d7a57f4d7de18fb0f92501172472c9b6",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Créer un nouvel utilisateur",
                "operationId": "9fea6f2adcb69614c521a7733358382d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "email",
                                    "role",
                                    "password"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "Client"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "123 rue de Paris"
                                    },
                                    "zone_intervention": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "ajoute_par": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "team": {
                                        "type": "string",
                                        "example": "Equipe 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Afficher un utilisateur spécifique",
                "operationId": "40a5104e23b52137e78c3d3ac48b88a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Mettre à jour un utilisateur",
                "operationId": "b166772cec9714067040e77cf14beeba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "John Smith"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "Superviseur"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "987654321"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "456 rue de Paris"
                                    },
                                    "zone_intervention": {
                                        "type": "string",
                                        "example": "Lyon"
                                    },
                                    "disponibilite": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "ce06860a51ab016ef66e753544fe4563",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé avec succès"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        },
        "/api/users/{id}/commandes": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Afficher les commandes d'un utilisateur",
                "operationId": "7589f960f9496d864b8f7f3a2262c2d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des commandes de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "Utilisateur",
                "required": [
                    "nom",
                    "email",
                    "role",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "Client"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "123456789"
                    },
                    "adresse": {
                        "type": "string",
                        "example": "123 rue de Paris"
                    },
                    "zone_intervention": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "disponibilite": {
                        "type": "boolean",
                        "example": true
                    },
                    "ajoute_par": {
                        "type": "string",
                        "example": "admin"
                    },
                    "team": {
                        "type": "string",
                        "example": "Equipe 1"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}